# contract/openapi.yaml
openapi: 3.1.0
info:
  title: Actual Bridge API
  version: 1.0.0
  description: |
    Stable contract between the Android app and the Actual bridge.

    Conventions
    - All money values are integers in minor units (cents). Ledger amounts are signed:
      outflows negative, inflows positive. Budget amounts (`budgeted`, `totalBudgeted`)
      are the amount assigned, positive when money is assigned.
    - Months are `YYYY-MM`, dates are `YYYY-MM-DD`.
    - Within /v1 changes are additive only. Clients must ignore unknown fields and
      treat unknown enum values as `unknown`.
    - Non-idempotent writes require an `Idempotency-Key` header. Creates use a
      client-generated UUID `id` instead.
    - Errors are `application/problem+json` with a stable `code`.
servers:
  - url: https://budget-api.example.com
security:
  - bearer: []
...
paths:
...
  /v1/setup/claim:
    post:
      tags: [setup]
      operationId: claimBridge
      summary: Become the first owner with the one-time setup code, signing the bridge in to Actual on the way.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetupClaim' }
      responses:
        '200':
          description: Owner created and this device signed in
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
        '503': { $ref: '#/components/responses/Problem' }
  /v1/auth/pair:
    post:
      tags: [auth]
      operationId: pairDevice
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PairRequest' }
      responses:
        '200':
          description: Device paired
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '400': { $ref: '#/components/responses/Problem' }
        '401': { $ref: '#/components/responses/Problem' }
  /v1/auth/refresh:
    post:
      tags: [auth]
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: New token pair (refresh tokens rotate)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
...
  responses:
    Problem:
      description: Error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    SetupClaim:
      type: object
      required: [setupCode, displayName, deviceName, platform]
      additionalProperties: false
      properties:
        setupCode: { type: string, minLength: 4, maxLength: 64, description: From the bridge logs or BRIDGE_SETUP_CODE }
        displayName: { type: string, minLength: 1, maxLength: 80 }
        deviceName: { type: string, minLength: 1, maxLength: 80 }
        platform: { type: string, enum: [android, ios, web, other] }
        actualPassword: { type: string, maxLength: 256, description: Required when actual is needs-password or needs-login }
...
    Role:
      type: string
      enum: [owner, member, viewer]
    Problem:
      type: object
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code:
          type: string
          description: Stable machine-readable code, e.g. `unauthorized`, `forbidden`, `not_found`, `validation`, `feature_unavailable`, `actual_unavailable`, `budget_encrypted`, `bank_sync_failed`, `update_failed`, `conflict`, `rate_limited`
        detail: { type: string }
...
    PairRequest:
      type: object
      required: [code, deviceName, platform]
      properties:
        code: { type: string, minLength: 6 }
        deviceName: { type: string, minLength: 1, maxLength: 80 }
        platform: { type: string, enum: [android, ios, web, other] }
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken: { type: string }
    TokenResponse:
      type: object
      required: [accessToken, refreshToken, expiresIn, member, device]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer, description: Access token lifetime in seconds }
        member: { $ref: '#/components/schemas/Member' }
        device: { $ref: '#/components/schemas/Device' }
    Member:
      type: object
      required: [id, displayName, role, disabled, budgetIds]
      properties:
        id: { type: string }
        displayName: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        disabled: { type: boolean }
        budgetIds:
          type: array
          description: Budgets this member may open. Owners implicitly have all.
          items: { type: string }
...
    Device:
      type: object
      required: [id, name, platform, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        platform: { type: string }
        createdAt: { type: string, format: date-time }
        lastSeenAt: { type: [string, 'null'], format: date-time }
...